Ignoring error messages
The gdal version that is distributed across all arch linux version,
apparently as linked against all possible optional dependencies
(mariadb, postgresdb, ...) and therefore gdal throws with every
process/interaction an error for each .so file that is missing.

Therefore to keep my sanity when using this python package, this version
ignores stderr and redirects it to /dev/null when actually creating tile
data.